Add aws lambda code
The project is now served from aws lambda. This allows
the server to bake the first set of words into the page so that
no external requests are needed. This, and the removal of the code
requied to parse the randomtext.org response results in a reduction of
about 2.3kB on the first request and 1kB on subsequent wordlist requests.
Lambdas are a bit copy paste and could be handled better tho.